+++ qaoa_vrp/generators/generator_utils.py
@@ -75,14 +75,6 @@ def instance_constructor(
     return instance
 
 
-def get_direction():
-    direction = np.random.randint(-1, 2)
-    if direction == 0:
-        return get_direction()
-    else:
-        return direction
-
-
 def compile_and_write(
     num_nodes,
     num_vehicles,
